+++ src/komodo_utils.cpp
@@ -761,37 +761,6 @@ int64_t komodo_block_unlocktime(uint32_t nHeight)
     return ((int64_t)unlocktime);
 }
 
-long _stripwhite(char *buf,int accept)
-{
-    int32_t i,j,c;
-    if ( buf == 0 || buf[0] == 0 )
-        return(0);
-    for (i=j=0; buf[i]!=0; i++)
-    {
-        buf[j] = c = buf[i];
-        if ( c == accept || (c != ' ' && c != '\n' && c != '\r' && c != '\t' && c != '\b') )
-            j++;
-    }
-    buf[j] = 0;
-    return(j);
-}
-
-char *clonestr(char *str)
-{
-    char *clone;
-    if ( str == 0 || str[0] == 0 )
-    {
-        printf("warning cloning nullstr.%p\n",str);
-#ifdef __APPLE__
-        while ( 1 ) sleep(1);
-#endif
-        str = (char *)"<nullstr>";
-    }
-    clone = (char *)malloc(strlen(str)+16);
-    strcpy(clone,str);
-    return(clone);
-}
-
 int32_t safecopy(char *dest,char *src,long len)
 {
     int32_t i = -1;
